Cleanup: make f_string_eol and f_string_placeholder a string instead of a char
In the back of my mind I keep reading f_string_eol as a string.
It is not.
Instead, it is a character.

To prevent this potential confusion, just make it a string and require the use of f_string_eol[0].
Make the same kind of change to f_string_placeholder as well.