Progress: featureless make
Handle the script sources differently than the bash sources.
The bash sources, a scrript, is for generated/processed (and the how is yet to be defined).
The more general scripts are meant for already prepared scripts not modified by this program (unless modified by some pre-process or post-process script).
The more general scripts does not care what the file is, it just gets copied over into the build scripts program path.

Add build_script yes/no setting to conditional include the general scripts when building.

Slightly rework the logic when processing the build operation functions.
Add the planned functions as stubs, to be completed in a future commit.