Bugfix: do not embed libraries into program
The generate.sh script design seems to be adding the library sources into the program.
This effectively makes the linked library.
This is a bad compiling and linking design and a waste of space.

Instead, add a link option to the project name so that the library gets linked to by the program.