Bugfix: invalid memory access and only process if not empty
If the path_headers is not defined, an invalid memory read was triggered.
This happens because I was only assigning path_headers variable when path_heades setting was defined but I was still attempting to copy that path regardless.

Make sure to not bother attempting any copy or compile build operations if that setting has not properties (such as build_sources_program being empty).