Update: Featureless Make needs to handle the -d/--define parameter
Rename the parameter from defines to define, because it acts on only a single parameter at a time.
The -d must be specified multiple times to get multiple defines.

When -d/--define is specified, the defines provided by the settings file get ignored.

Continue to pass the defines from the console parameters to the process scripts.