Update: bootstrap dependency functionality, fix missing dependencies, also ensure level and monolithic
It seems I wrote a dependency generator to make my life easier and then I went and forgot about it.
Update this to work with the more recent changes to the build settings structure.

While I am at it, generate all of the dependencies and save the generated dependencies.

This exposed a typoe in the dependencies file, now fixed.

Might as well make sure the level and monolithic dependencies are generated as well.