Update: fl_serialized to f_serialize, renaming f_serialized to f_serialize
There is nothing forcing the current functions in fl_serialized from being in f_serialize.

Renamed f_serialized to f_serialize to be more consistent with the function names.
To that end, I have renamed the functions like f_unserialize* to f_serialize_un.
I have been unable to come up with a better name that is also reasonably short.