Feature: add fss_named and fss_set structures and re-organize the fss headers
Provide fss_named and fss_set structures (and their respective array structures).
These are useful when using FSS types nested within another set (not to be confused with fss_nest which is for a single type with recursion).

There are just too many things in fss.h.
Break it out into multiple headers that are auto-included.