Update: UTF-8 improvements
Move some FSS UTF-8 processing code into UTF-8 as it should be more generic.

Add more UTF-8 functions (or at least the stubs).
Begin adding emoji, which is all over the place.

Staring at so many codes, it will be easy to make a mistake.
After all of the UTF-8 functions are believed to be fully implemented an extensive review needs to happen.