Progress: featureless make.
Begin working on "if" operation logic.

Make sure the return states are handled by adding a "success" boolean.

Uncomment the code where I was confused by the logic.
What seems to have caused the confusion was that the "if" condition should have been an "else if" condition.
With this change, the logic makes more sense because it is no longer checking against a modified status variable.