Update: console improvements and fixes.
Get rid of total property.
This is redundant and too short.
Instead, use the location.used, which is an unsigned long as opposed to a uint8_t.

When I changed the result from an f_string_length_t to an f_array_length_t, I missed some places where the type needed to be replaced.