Feature: add support for trimming FSS Object names for FSS write operations.
I believe it is important for the command line tools to easily cleanup object names in case spaces aren't desired.
This provides the functionality to ensure trimming of spaces before and after the object name.

This further allows for preventing the cleanup behavior done in object writes, such as in FSS-0003 (Extended List) where it is cleaner to have a space between the object name and the brace '{'.
If this is undesired, the function can then be configured to not do this.

This further makes the FSS write functions more consistent with the FSS read functions where object name trimming is already supported.