Bugfix: trimming in FSS extended write does not always work.
A boolean is used to designate when the Object has no space at the end.
If this boolean is FALSE, then a space is added before the '{'.
The problem is that the addition of this space happens after trimming.

If trimming is performed, just lie and set the boolean to TRUE to prevent the addition of an (unwanted) space.