Update: FSS should allow caller to determine when to apply delimits.
Instead of automatically applying the delimits, pass the delimits to the caller.
Let the caller decide when or when not to apply the delimits.
Provide functions to help in this regard.

The FSS read programs will need a new parameter to allow for designating to or not to apply delimits.
This is particularly complex when it comes to FSS-0003 (Extended List) due to the nesting.

This change gave me the opportunity to review the FSS read functions, to a certain extent.
Code cleanup changes are included in this commit.

Also add status codes: F_rrange, F_range_not, and F_utf_not.