Bugfix: consistently use the data.output file for standard output printing.
There are still places that need to be converted from f_type_output and printf() to use this data.output file.
With this change, the data.output will be used and can more readily be updated to point to some file other than standard output.

The fss_*_read programs are not updated with this commit due to having additional changes.
These changes will also be applied to those programs in another commit.