Update: FSS Extended List should now support multiple depth parameters.
Recursively parses the parameters determining what data should or should not be processed.

For example, take the following file "example.fss":
a {
  a.0 {
    a.0.0 {
     a's final content.
    }
  }
}
b {
  b.0 {
   b's first final content.
  }
  b.1 {
   b's second final content.
  }
}

TO only display "b's first final content.", the following is done:
  fss_extended_list_read example.fss -n 'b ' -d 1 -a 0

Which works by at depth 0, looks for the list name 'b ', then at depth 1, looks at position 0.

The --at and --name parameters are using the wrong variable in their logic (using 'j' when it should be 'i').
This resulted in an incorrect parsing of the parameters.