Bugfix: dependency generation is incorrect.
The recent rewrite of the dependencies resulted in the higher level files being after the lower level files.
Normally this is not a problem, but when it comes to compiling static programs depending on this, the linking fails.
This becomes more confusing in that, in my case the GNU Linker, says that functions do not exist when you can see that they, in fact, do exist.

By changing the order, with the higher levels before the lower levels, static compilation works without any such issues.