Bugfix: color print is bleeding into GCC messages.
When GCC has an error or warning, the print messages end up discolored.

This seems to be an artifact of how streams work and the newline.
Move all final newlines out of the color context printing and print the newlines in a separate command.
This appears to avoid the mentioned problem.